Move the test Co-authored-by: Alex Pope <popey2700@users.noreply.github.com>
